mini-game-automation/requirements.txt
# Core dependencies
playwright>=1.40.0
opencv-python>=4.8.0
pillow>=10.0.0
easyocr>=1.7.0
portalocker>=2.8.0

# System monitoring
psutil>=5.9.0

# YAML configuration
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Type checking (optional)
# mypy>=1.7.0

# Code formatting (optional)
# black>=23.0.0
# flake8>=6.1.0
